+++ packages/vitepress-plugin-image-preview/CHANGELOG.md
@@ -1,7 +1,7 @@
-# vitepress-plugin-slot-inject-template
+# vitepress-plugin-image-preview
 
 ## 0.1.0
 
 ### Minor Changes
 
-- feat: init plugin
+- feat: split-image-preview